[DF-5671] Fixed cast warnings
Former-commit-id: 09d1ca8048e431b3e637183f1ac24a936ee12046

+++ Sources/Internal/Scene3D/Entity.cpp
@@ -153,7 +153,7 @@ void Entity::AddComponent(Component * component)
 #endif
 	}
 		
-	componentFlags |= (1 << (uint64)component->GetType());
+	componentFlags |= ((uint64)1 << (uint64)component->GetType());
 
 	if (scene)
 		scene->RegisterComponent(this, component);
